refactor: CardList 컴포넌트 components 폴더로 이동
